Changed emails for Capital invest, added translations and all working
