+++ contact/templates/email/capital_invest_email_agent.txt
@@ -1,0 +1,27 @@
+{% load trans from i18n %}
+
+This is confirmation of a Capital Invest lead via the contact us form on the website.
+
+See below for the user submitted form:
+
+{% for name, value in form_data %}
+    {{ name }}
+    {{ value }}
+{% endfor %}
+
+{% trans "Submission url" %}
+{{ request.path }}
+
+Google campaign:
+{% for utm_code, utm_value in utm.items %}
+    {{ utm_code }}: {{ utm_value }}<br>
+{% empty %}
+    None.
+{% endfor %}
+
+
+Many Thanks
+DIT
+
+Department for International Trade (DIT) is the Government Department that helps UK‑based companies succeed in the global economy. We also help overseas companies bring their high-quality investment to the UK’s dynamic economy,
+acknowledged as Europe’s best place in which to succeed in global business.https://invest.great.gov.uk
