+++ contact/templates/email/capital_invest_email_user.txt
@@ -1,0 +1,21 @@
+Invest in GREAT Britain Contact Confirmation
+
+Thank you for contacting the Department for International Trade.
+
+We have received the information you sent through the Capital Invest website. Your enquiry may be forwarded to a local post for follow up.
+
+The Department for International Trade provides free and impartial advice to companies around the world interested in doing business in the UK. We look forward to welcoming you as one of the many companies to enjoy success in the UK.
+
+Find our Terms and Conditions at https://great.gov.uk/terms-and-conditions
+
+See below for your submitted form:
+{% for name, value in form_data %}
+	{{ name }}
+	{{ value }}
+{% endfor %}
+
+Many Thanks
+DIT
+
+Department for International Trade (DIT) is the Government Department that helps UK‑based companies succeed in the global economy. We also help overseas companies bring their high-quality investment to the UK’s dynamic economy,
+acknowledged as Europe’s best place in which to succeed in global business.https://invest.great.gov.uk
