+++ contact/views.py
@@ -8,7 +8,6 @@ from contact import forms
 from contact.mixins import LocalisedURLsMixin
 
 from directory_components.mixins import EnableTranslationsMixin, InternationalHeaderMixin
-from core.mixins import CMSPageFromSlugMixin
 
 
 class ContactFormView(
@@ -59,62 +58,6 @@ class ContactFormSuccessView(
         self.set_ga360_payload(
             page_id='InvestContactFormSuccess',
             business_unit='Invest',
-            site_section='Contact',
-            site_subsection='ContactSuccess'
-        )
-
-
-class CapitalInvestContactFormView(
-    CMSPageFromSlugMixin,
-    LocalisedURLsMixin,
-    InternationalHeaderMixin,
-    EnableTranslationsMixin,
-    CountryDisplayMixin,
-    GA360Mixin,
-    FormView,
-):
-    success_url = reverse_lazy('capital-invest-contact-success')
-    slug = 'contact'
-    page_type = 'CapitalInvestContactFormPage'
-    template_name = 'core/capital_invest/capital_invest_contact_form.html'
-    form_class = forms.CapitalInvestContactForm
-    available_languages = settings.LANGUAGES
-
-    def __init__(self):
-        super().__init__()
-        self.set_ga360_payload(
-            page_id='CapitalInvestContactForm',
-            business_unit='CapitalInvest',
-            site_section='Contact'
-        )
-
-    def get_form_kwargs(self):
-        kwargs = super().get_form_kwargs()
-        kwargs['utm_data'] = self.request.utm
-        kwargs['submission_url'] = self.request.path
-        return kwargs
-
-    def form_valid(self, form):
-        form.save()
-        return super().form_valid(form)
-
-
-class CapitalInvestContactFormSuccessView(
-    LocalisedURLsMixin,
-    InternationalHeaderMixin,
-    EnableTranslationsMixin,
-    CountryDisplayMixin,
-    GA360Mixin,
-    TemplateView,
-):
-    template_name = 'core/capital_invest/capital_invest_contact_form_success.html'
-    available_languages = settings.LANGUAGES
-
-    def __init__(self):
-        super().__init__()
-        self.set_ga360_payload(
-            page_id='CapitalInvestContactFormSuccess',
-            business_unit='CapitalInvest',
             site_section='Contact',
             site_subsection='ContactSuccess'
         )
