+++ core/templates/core/capital_invest/capital_invest_contact_form.html
@@ -13,14 +13,17 @@
 {% endblock %}
 
 {% block content %}
-{% block breadcrumbs %}
-    <div class="container">
-        {% breadcrumbs page.breadcrumbs_label %}
-            <a href="{{ site_home_link.url }}">{{ site_home_link.label }}</a>
-            <a href="international/content/capital-invest">Capital Invest</a>
-        {% endbreadcrumbs %}
-    </div>
-{% endblock %}
+
+{% if page.breadcrumbs_label %}
+    {% block breadcrumbs %}
+        <div class="container">
+            {% breadcrumbs page.breadcrumbs_label %}
+                <a href="{{ site_home_link.url }}">{{ site_home_link.label }}</a>
+                <a href="international/content/capital-invest">Capital Invest</a>
+            {% endbreadcrumbs %}
+        </div>
+    {% endblock %}
+{% endif %}
 
 <div class="container">
     {% ga360_data "contact-form" target="form" %}
