+++ core/templates/core/capital_invest/capital_invest_contact_form_success.html
@@ -1,5 +1,5 @@
 {% extends 'contact/base.html' %}
-{% load breadcrumbs from directory_components %}
+{% load breadcrumbs add_export_elements_classes from directory_components %}
 {% load i18n %}
 
 {% block content %}
@@ -8,8 +8,8 @@
 <div class="container">
   <div class="grid-row">
     <div class="column-two-thirds">
-      <h2 class="heading-large">{% trans "Your form has been submitted." %}</h2>
-      <p>{% trans "Thank you, your form has been submitted." %}</p>
+      <h2 class="heading-large">{{ page.large_text }}</h2>
+      <p>{{ page.small_text|add_export_elements_classes|safe }}</p>
     </div>
   </div>
 </div>
