+++ locale/pt/LC_MESSAGES/django.po
@@ -8,7 +8,7 @@ msgid ""
 msgstr ""
 "Project-Id-Version: PACKAGE VERSION\n"
 "Report-Msgid-Bugs-To: \n"
-"POT-Creation-Date: 2019-08-06 14:02+0000\n"
+"POT-Creation-Date: 2019-08-08 14:11+0000\n"
 "PO-Revision-Date: YEAR-MO-DA HO:MI+ZONE\n"
 "Last-Translator: FULL NAME <EMAIL@ADDRESS>\n"
 "Language-Team: LANGUAGE <LL@li.org>\n"
@@ -18,51 +18,51 @@ msgstr ""
 "Content-Transfer-Encoding: 8bit\n"
 "Plural-Forms: nplurals=2; plural=(n != 1);\n"
 
-#: contact/forms.py:19
+#: contact/forms.py:18
 msgid "Less than 10"
 msgstr "Menos de 10"
 
-#: contact/forms.py:23
+#: contact/forms.py:22
 msgid "10 to 50"
 msgstr "Entre 10 e 50"
 
-#: contact/forms.py:27
+#: contact/forms.py:26
 msgid "51 to 250"
 msgstr "Entre 51 e 250"
 
-#: contact/forms.py:31
+#: contact/forms.py:30
 msgid "More than 250"
 msgstr "Mais de 250"
 
-#: contact/forms.py:39 contact/forms.py:87 perfect_fit_prospectus/forms.py:12
+#: contact/forms.py:38 contact/forms.py:87 perfect_fit_prospectus/forms.py:12
 msgid "Name"
 msgstr "Nome completo"
 
-#: contact/forms.py:40 contact/forms.py:89
+#: contact/forms.py:39 contact/forms.py:89
 msgid "Job title"
 msgstr "Cargo"
 
-#: contact/forms.py:41 contact/forms.py:88
+#: contact/forms.py:40 contact/forms.py:88 core/forms.py:99 core/forms.py:136
 msgid "Email address"
 msgstr "E-mail de trabalho"
 
-#: contact/forms.py:43 contact/forms.py:90
+#: contact/forms.py:42 contact/forms.py:90
 msgid "Phone number"
 msgstr "Telefone comercial"
 
-#: contact/forms.py:46 contact/forms.py:91
+#: contact/forms.py:45 contact/forms.py:91
 msgid "Company name"
 msgstr "Nome da empresa registado"
 
-#: contact/forms.py:48 contact/forms.py:92
+#: contact/forms.py:47 contact/forms.py:92
 msgid "Company website"
 msgstr "Site da empresa"
 
-#: contact/forms.py:52
+#: contact/forms.py:51 core/forms.py:101
 msgid "Which country are you based in?"
 msgstr "País"
 
-#: contact/forms.py:54
+#: contact/forms.py:53 core/forms.py:103
 msgid ""
 "We will use this information to put you in touch with your closest British "
 "embassy or high commission."
@@ -71,15 +71,15 @@ msgstr ""
 "britânica ou representante mais próximo. Comece a digitar para selecionar "
 "uum país."
 
-#: contact/forms.py:60 contact/forms.py:94
+#: contact/forms.py:59 contact/forms.py:94
 msgid "Current number of staff"
 msgstr "Número total de funcionários atual"
 
-#: contact/forms.py:64
+#: contact/forms.py:63
 msgid "Tell us about your investment"
 msgstr "Conte-nos sobre o seu investimento"
 
-#: contact/forms.py:66
+#: contact/forms.py:65
 msgid ""
 "Tell us about your company and your plans for the UK in terms of size of "
 "investment, operational and recruitment plans. Please also tell us what help "
@@ -90,7 +90,7 @@ msgstr ""
 "Explique-nos que tipo de ajuda você gostaria de receber do governo do Reino "
 "Unido."
 
-#: contact/forms.py:93 perfect_fit_prospectus/forms.py:32
+#: contact/forms.py:93 core/forms.py:137 perfect_fit_prospectus/forms.py:32
 msgid "Country"
 msgstr "País"
 
@@ -98,7 +98,7 @@ msgstr "País"
 msgid "Your investment"
 msgstr "Seu investimento"
 
-#: contact/forms.py:127
+#: contact/forms.py:127 core/forms.py:171
 msgid "Contact form user email subject"
 msgstr ""
 
@@ -112,6 +112,7 @@ msgid "Contact the investment team"
 msgstr "Contacte a nossa equipa de investimento"
 
 #: contact/templates/contact/contact.html:24
+#: core/templates/core/capital_invest/capital_invest_contact_form.html:38
 msgid "Contact Information"
 msgstr "Informações de contato"
 
@@ -152,6 +153,8 @@ msgstr ""
 "recebemos as suas informações. Pretendemos enviar-lhe um e-mail de "
 "seguimento dentro de 7 dias. Estamos ansiosos por trabalhar consigo."
 
+#: contact/templates/email/capital_invest_email_agent.html:166
+#: contact/templates/email/capital_invest_email_agent.txt:12
 #: contact/templates/email/email_agent.html:166
 #: contact/templates/email/email_agent.txt:12
 msgid "Submission url"
@@ -161,24 +164,50 @@ msgstr ""
 msgid "Great.gov.uk International"
 msgstr ""
 
-#: core/forms.py:22
+#: core/forms.py:34
 msgid "sector"
 msgstr ""
 
-#: core/forms.py:30
+#: core/forms.py:42
 msgid "scale"
 msgstr ""
 
-#: core/forms.py:38
+#: core/forms.py:50
 msgid "region"
 msgstr ""
 
-#: core/forms.py:46
+#: core/forms.py:58
 msgid "sub_sector"
 msgstr ""
 
-#: core/forms.py:54
+#: core/forms.py:66
 msgid "sort_by"
+msgstr ""
+
+#: core/forms.py:97
+msgid "Given name"
+msgstr ""
+
+#: core/forms.py:98
+#, fuzzy
+#| msgid "Company name"
+msgid "Family name"
+msgstr "Nome da empresa registado"
+
+#: core/forms.py:108 core/forms.py:138
+msgid "City"
+msgstr ""
+
+#: core/forms.py:110 core/forms.py:139
+msgid "Message"
+msgstr ""
+
+#: core/forms.py:134
+msgid "Given Name"
+msgstr ""
+
+#: core/forms.py:135
+msgid "Family Name"
 msgstr ""
 
 #: core/templates/404.html:12
