+++ makefile
@@ -37,7 +37,8 @@ DEBUG_SET_ENV_VARS := \
 	export CMS_URL=http://cms.trade.great:8010; \
 	export CMS_SIGNATURE_SECRET=debug; \
 	export DEFAULT_FROM_EMAIL=jessica.gilbert@digital.trade.gov.uk; \
-	export IIGB_AGENT_EMAIL=j.jgilbert@yahoo.com; \
+	export IIGB_AGENT_EMAIL=jessica.gilbert@digital.trade.gov.uk; \
+	export CIGB_AGENT_EMAIL=jessica.gilbert@digital.trade.gov.uk; \
 	export EMAIL_HOST=digital.trade.gov.uk; \
 	export EMAIL_HOST_USER=debug; \
 	export EMAIL_HOST_PASSWORD=debug; \
